Created a 'nav' folder which contains a 'components' folder and a 'Pages' folder. In the 'pages' folder, created jsx & css files for History, Home & Themes. Moved burger menu.jsx & .css into nav/components.
