+++ src/App.css
@@ -1,7 +1,7 @@
 .app {
   display: grid;
   height: 100vh;
-  width: 100%;
+  width: 100%vw;
   background-color: #2d2d86;
 }
 
@@ -14,9 +14,9 @@
   grid-template-columns: 1fr 1fr 1fr;
   grid-template-rows: 1fr 1fr 1fr;
   grid-template-areas:
-  "header header header"
-  "calc-border-left calc-wrapper calc-border-right"
-  "footer footer footer"
+  "header            header        header"
+  "calc-border-left  calc-wrapper  calc-border-right"
+  "footer            footer        footer";
 }
 
 .calc-border-left {
