--- src/components/hamburger-menu/Themes.css
@@ -1,3 +1,0 @@
-.theme-wrapper {
-    background-color: darksalmon;
-}
